Changed the argument order when calling an app with data
This breaks symetry with the call_app_with_data, which adds the data
to the end of the args list: when calling the app, the data arg is
used *before* Env. Maybe counter intertuitive. The rationale is that,
from the caller's standpoint, Data is the extra item in the call, and
should therefore be at the end of the list. For the app however, more
specific variants of "handle app" go *before* the common state Env.
This follows the convention used in gen_server, where State is always
the last argument -- and more specific arguments are listed in
front of State. Confusing? Maybe.