Very remedial support for server callback module
This facility can be used to perform init operations for a server
process (e.g. register a server, monitor other processes, etc.)

It can also be used to handle messages (e.g. sent to the process
directly or via proc:{cast,call}.

Note this is not facility for high througput interactions with a
server - Psycho servers are still very much dedicated to handling
inbound requests as quickly (and naively) as possible. Messages
must wait for either a new connection or a timeout (specified by
the accept_timeout server option). This is suitable for trivial
message handling like stopping the server or casually swapping out
the root app for the next request. I.e. messages are handled in
due time, being delayed at most by the accept_timeout.