v2 static support
This is a much better interface, using options that can be updated
without breaking the API.

The driver was to support etags and 304 (not modified) responses.

Static apps can be configured with a default cache control. This isn't
as flexible as one might want as the policy applies to all content
served by the app. Of course different apps with different policies
can be used - each applied to a different directory. But this still
isn't as flexible as it might be - where policies can be applied based
on rules, or even with callbacks.